requirements.txt
numpy
discord==1.7.3
discord.py==1.7.3
pyshorteners
schedule
python-dotenv